+++ data_service/card_order_change_log_m.go
@@ -1,13 +1,15 @@
 package data_service
 
 type CardOrderChangeLogInfo struct {
-	Id          int64  `json:"id" gorm:"column:id"`
-	CardOrderId int32  `json:"card_order_id" gorm:"column:card_order_id"`
-	ChangeLog   string `json:"change_log" gorm:"column:change_log"`
-	Describe    string `json:"describe" gorm:"column:describe"`
-	Mtime       uint32 `json:"mtime" gorm:"autoUpdateTime <-:update"`
-	Ctime       uint32 `json:"ctime" gorm:"autoCreateTime <-:create"`
-	Deleted     int32  `json:"deleted" gorm:"column:deleted"`
+
+	Id           uint64  `json:"id" gorm:"column:id"`
+	CardOrderId	uint64	`json:"card_order_id" gorm:"column:card_order_id"`
+	ChangeLog	string	`json:"change_log" gorm:"column:change_log"`
+	Describe	string	`json:"describe" gorm:"column:describe"`
+	Mtime   uint32  `json:"mtime" gorm:"autoUpdateTime <-:update"`
+	Ctime   uint32  `json:"ctime" gorm:"autoCreateTime <-:create"`
+	Deleted int32  `json:"deleted" gorm:"column:deleted"`
+
 }
 
 func (ui *CardOrderChangeLogInfo) TableName() string {
